remove view in js
