+++ js/templates/change-state-game.js
@@ -1,5 +1,5 @@
-import GetStateGame from "./view/stateGameView";
-import {renderTemplate} from "../utils";
+import GetStateGame from "./../view/stateGameView";
+import {renderTemplate} from "./../utils";
 
 const renderState = (state) => {
   const elementStateGame = new GetStateGame(state);
