+++ js/templates/succes-module.js
@@ -1,7 +1,7 @@
 import {renderTemplate} from "./../utils";
 import welcomeScreen from "./welcome-screen";
 import {results} from "./../data/models/endGame";
-import SuccessView from "./view/result-success-module";
+import SuccessView from "./../view/result-success-module";
 
 const renderSuccess = () => {
   const successView = new SuccessView(results);
