+++ js/templates/time-out-screen.js
@@ -1,7 +1,7 @@
 import {renderTemplate} from "./../utils";
 import welcomeScreen from "./welcome-screen";
-import {results} from "../data/models/endGame";
-import TimeOutView from "./view/timeOutView";
+import {results} from "./../data/models/endGame";
+import TimeOutView from "./../view/timeOutView";
 
 
 const timeOutView = new TimeOutView(results);
