+++ js/templates/welcome-screen.js
@@ -1,8 +1,8 @@
-import WelcomeView from "./../templates/view/welcomeView";
+import WelcomeView from "./../view/welcomeView";
 import {startGame} from "./../data/models/startGame";
 import {renderTemplate} from "./../utils";
-import {initializationState} from "../data/models/initialization-state";
-import changeView from "../changeState";
+import {initializationState} from "./../data/models/initialization-state";
+import changeView from "./../changeState";
 
 const renderWelcome = () => {
   const previewView = new WelcomeView(startGame);
