+++ js/view/artistView.js
@@ -1,4 +1,4 @@
-import AbstractView from "../../abstract-view";
+import AbstractView from "./../abstract-view";
 
 class ArtistView extends AbstractView {
   constructor(data, level) {
