+++ js/view/genreView.js
@@ -1,4 +1,4 @@
-import AbstractView from "./../../abstract-view";
+import AbstractView from "./../abstract-view";
 
 class GenreView extends AbstractView {
   constructor(data, level) {
