+++ js/view/result-success-module.js
@@ -1,4 +1,4 @@
-import AbstractView from "./../../abstract-view";
+import AbstractView from "./../abstract-view";
 
 
 class SuccessView extends AbstractView {
