+++ js/view/stateGameView.js
@@ -1,4 +1,4 @@
-import AbstractView from "./../../abstract-view";
+import AbstractView from "./../abstract-view";
 
 class GetStateGame extends AbstractView {
   constructor(data) {
@@ -15,7 +15,7 @@ class GetStateGame extends AbstractView {
       <circle
         cx="390" cy="390" r="370"
         class="timer-line"
-        style="filter: url(../..#blur); transform: rotate(-90deg) scaleY(-1); transform-origin: center"></circle>
+        style="filter: url(..#blur); transform: rotate(-90deg) scaleY(-1); transform-origin: center"></circle>
       <div class="timer-value" xmlns="http://www.w3.org/1999/xhtml">
         <span class="timer-value-mins">05</span>
         <span class="timer-value-dots">:</span>
