+++ js/view/timeOutView.js
@@ -1,7 +1,7 @@
-import AbstractView from "./../../abstract-view";
+import AbstractView from "./../abstract-view";
 
 
-class AttemptsEnded extends AbstractView {
+class TimeOutView extends AbstractView {
   constructor(data) {
     super();
     this.data = data;
@@ -10,8 +10,8 @@ class AttemptsEnded extends AbstractView {
   get template() {
     return `<section class="main main--result">
     <section class="logo" title="${this.data.title}"><h1>${this.data.title}</h1></section>
-    <h2 class="title">${this.data.attempsEnded.title}</h2>
-    <div class="main-stat">${this.data.attempsEnded.endsOfAttempts}</div>
+    <h2 class="title">${this.data.timeOut.title}</h2>
+    <div class="main-stat">${this.data.timeOut.endsOfTime}</div>
     <span role="button" tabindex="0" class="main-replay">${this.data.playAgainLose}</span>
   </section>`;
   }
@@ -26,4 +26,4 @@ class AttemptsEnded extends AbstractView {
   }
 }
 
-export default AttemptsEnded;
+export default TimeOutView;
