+++ js/view/welcomeView.js
@@ -1,4 +1,4 @@
-import AbstractView from "../../abstract-view";
+import AbstractView from "./../abstract-view";
 
 class WelcomeView extends AbstractView {
   constructor(data) {
